Revamp CLI with captcha window
