+++ README.md
@@ -1,16 +1,16 @@
 # Инструкция по запуску fsrar_cli.py
 
-Эта утилита позволяет отправить запрос на https://check1.fsrar.ru/ и получить
-информацию о статусе накладной.
+Эта утилита отправляет запрос на https://check1.fsrar.ru/ и
+получает информацию о статусе накладной.
 
 ## Требования
 - Python 3.8+
-- Установленные библиотеки `requests` и `beautifulsoup4`
+- Установленные библиотеки `requests`, `beautifulsoup4` и `pillow`
 
 Установить зависимости можно командой:
 
 ```bash
-pip install requests beautifulsoup4
+pip install requests beautifulsoup4 pillow
 ```
 
 ## Запуск
@@ -20,8 +20,8 @@ pip install requests beautifulsoup4
    python fsrar_cli.py
    ```
 2. Введите номер ТТН и ФСРАР ID получателя.
-3. Откройте в браузере ссылку на изображение капчи,
-   введите капчу и дождитесь ответа от сервера.
+3. После получения капчи откроется небольшое окно с её изображением.
+   Введите символы и нажмите «Отправить».
 
 Ответ появится в консоли в формате JSON.
 
@@ -45,17 +45,3 @@ pip install requests beautifulsoup4
    pyinstaller --onefile fsrar_cli.py
    ```
 3. После сборки запустите полученный файл из директории `dist`.
-
-## Графический режим
-
-Для запуска программы с окном используйте скрипт `fsrar_gui.py`:
-
-```bash
-python fsrar_gui.py
-```
-
-1. Введите номер ТТН и ФСРАР ID получателя в появившемся окне и нажмите «Далее».
-2. Во втором окне откроется изображение капчи. Введите символы и нажмите «Отправить».
-3. Ответ от сервера отобразится во всплывающем сообщении.
-
-Файл журнала `fsrar_cli.log` сохраняет полное логирование и для графического режима.
